requirements.txt
Django==5.1.6
djangorestframework==3.15.2
requests==2.32.3
whitenoise==6.9.0
